ノート一覧ページの修正
