+++ next/src/components/common/Error.tsx
@@ -4,8 +4,12 @@ const Error = () => {
   return (
     <Container maxWidth="sm">
       <Card sx={{ p: 3, mt: 8, backgroundColor: '#EEEEEE' }}>
-        <CardContent sx={{ lineHeight: 2 }}>
-          現在、システムに技術的な問題が発生しています。ご不便をおかけして申し訳ありませんが、復旧までしばらくお待ちください。
+        <CardContent sx={{ lineHeight: 2, textAlign: 'center' }}>
+          現在、システムに技術的な問題が発生しています
+          <br />
+          ご不便をおかけして申し訳ありませんが、
+          <br />
+          復旧までしばらくお待ちください
         </CardContent>
       </Card>
     </Container>
