+++ next/src/components/note/CheerStatus.tsx
@@ -1,18 +1,17 @@
 import { Box } from '@mui/material'
 import Image from 'next/image'
 import { useRouter } from 'next/router'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuth } from '@/hooks/useAuth'
 import { useCheerStatus } from '@/hooks/useCheerStatus'
 import { useSnackbarState } from '@/hooks/useSnackbarState'
-import { handleError } from '@/requests/utils/handleError'
+import { handleError } from '@/utils/handleError'
 
-type CheerStatusProps = {
-  noteId: number
-}
-
-const CheerStatus = (props: CheerStatusProps) => {
-  const { idToken } = useAuth()
-  const { data, error } = useCheerStatus(props.noteId, idToken || undefined)
+const CheerStatus = ({ noteId }: { noteId: number }) => {
+  const { idToken, isAuthLoading } = useAuth()
+  const { data, error, isLoading } = useCheerStatus(
+    noteId,
+    idToken || undefined,
+  )
   const [, setSnackbar] = useSnackbarState()
   const router = useRouter()
 
@@ -35,7 +34,15 @@ const CheerStatus = (props: CheerStatusProps) => {
           height={16}
         />
       )}
-      {!data && <Box sx={{ width: 16 }} />}
+      {(isAuthLoading || isLoading) && <Box sx={{ width: 16 }} />}
+      {!isAuthLoading && !isLoading && !data && (
+        <Image
+          src="/megaphone-outlined.svg"
+          alt="Cheer Icon"
+          width={16}
+          height={16}
+        />
+      )}
       {data && (
         <Image
           src={
