+++ next/src/contexts/AuthContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react'
+import { AuthContextProps } from '@/types/auth'
+
+export const AuthContext = createContext<AuthContextProps | undefined>(
+  undefined,
+)
