+++ next/src/hooks/useCheerStatus.ts
@@ -1,11 +1,11 @@
 import useSWR, { SWRResponse } from 'swr'
-import { CheerResponse } from '@/requests/types/cheerResponse'
-import { fetcher } from '@/requests/utils/fetcher'
+import { CheerResponse } from '@/types/cheer'
+import { fetcher } from '@/utils/fetcher'
 
 export const useCheerStatus = (noteId: number, idToken?: string) => {
   const url = `${process.env.NEXT_PUBLIC_API_BASE_URL}/notes/${noteId}/cheer`
 
-  const { data, error }: SWRResponse<CheerResponse> = useSWR(
+  const { data, error, isLoading }: SWRResponse<CheerResponse> = useSWR(
     idToken && [url, idToken],
     fetcher,
   )
@@ -13,5 +13,6 @@ export const useCheerStatus = (noteId: number, idToken?: string) => {
   return {
     data,
     error,
+    isLoading,
   }
 }
