+++ next/src/hooks/useNotes.ts
@@ -1,16 +1,14 @@
 import useSWR, { SWRResponse } from 'swr'
-import { NotesResponse } from '@/requests/types/notesResponse'
-import { fetcher } from '@/requests/utils/fetcher'
+import { PagenatedNotesResponse } from '@/types/note'
+import { fetcher } from '@/utils/fetcher'
 
-export const useNotes = (page: number) => {
+export const useNotes = (page?: number) => {
   const url = page
     ? `${process.env.NEXT_PUBLIC_API_BASE_URL}/notes`
     : `${process.env.NEXT_PUBLIC_API_BASE_URL}/notes/?page=${page}`
 
-  const { data, error, isLoading }: SWRResponse<NotesResponse> = useSWR(
-    [url, undefined],
-    fetcher,
-  )
+  const { data, error, isLoading }: SWRResponse<PagenatedNotesResponse> =
+    useSWR([url, undefined], fetcher)
 
   return {
     data,
