+++ next/src/pages/auth/signup.tsx
@@ -12,7 +12,7 @@ import { useRouter } from 'next/router'
 import { useState } from 'react'
 import { useForm, SubmitHandler, Controller } from 'react-hook-form'
 import { useSnackbarState } from '@/hooks/useSnackbarState'
-import { handleError } from '@/requests/utils/handleError'
+import { handleError } from '@/utils/handleError'
 import { styles } from '@/styles'
 import auth from '@/utils/firebaseConfig'
