--- next/src/requests/types/cheerResponse.ts
@@ -1,3 +1,0 @@
-export type CheerResponse = {
-  hasCheered: boolean
-}
