--- next/src/requests/types/profileResponse.ts
@@ -1,11 +1,0 @@
-export type ProfileResponse = {
-  id: number
-  nickname?: string
-  bio?: string
-  xLink?: string
-  githubLink?: string
-  avatarUrl?: string
-  user: {
-    id: number
-  }
-}
