--- next/src/requests/types/relationshipResponse.ts
@@ -1,3 +1,0 @@
-export type RelationshipResponse = {
-  hasFollowed: boolean
-}
