--- next/src/requests/types/updateAuthPayload.ts
@@ -1,9 +1,0 @@
-export type UpdateAuthPayload = {
-  note: {
-    title?: string
-    content?: string
-    statusJp: '未保存' | '下書き' | '公開中'
-  }
-  duration: number
-  tags_names?: string[]
-}
