--- next/src/requests/types/updateNotePayload.ts
@@ -1,9 +1,0 @@
-export type UpdateNotePayload = {
-  note: {
-    title?: string
-    content?: string
-    statusJp: '未保存' | '下書き' | '公開中'
-  }
-  duration: number
-  tags_names?: string[]
-}
