--- next/src/requests/types/updateProfilePayload.ts
@@ -1,6 +1,0 @@
-export type UpdateProfilePayload = {
-  nickname?: string
-  bio?: string
-  xUsername?: string
-  githubUsername?: string
-}
