+++ next/src/types/auth.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react'
+
+export interface AuthContextProps {
+  idToken: string | null
+  isAuthLoading: boolean
+  setIsAuthLoading: React.Dispatch<React.SetStateAction<boolean>>
+}
+
+export interface AuthProviderProps {
+  children: ReactNode
+}
