+++ next/src/types/cheer.ts
@@ -1,0 +1,3 @@
+export interface CheerResponse {
+  hasCheered: boolean
+}
