+++ next/src/types/relationship.ts
@@ -1,0 +1,3 @@
+export interface RelationshipResponse {
+  hasFollowed: boolean
+}
